requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
numpy==2.2.5
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
pydantic==2.11.4
pydantic_core==2.33.2
pytest==8.3.5
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
